requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
